Fixed relative paths.
Problem: If one attempts to use this library as a part of another Arduino library, the sketch will not compile.
Reason: The source files in the \core and \platform contain header includes that do not have a path included, they point to header files within the same directory.
Solution: Add "../../" to header includes in \core and \platform, or,