Extend API - Async and Continuous
Slowely trying to add some of the additional capbilities of
these sensors to the top level library.

I am trying to make use of the underlying ST library that is included
with this library.

Some of the additions include:

The ability to startup a range measurement and not have to wait for it to complete.

The ability to start up a continue ranging operation

I included an example sketch that shows some different modes of
operation, it is setup with a table of sensors, like which SPI buss it
is on, which GPIO pins are for shutdown and GPIO...  Currently it is
doing 4 sensors 2 on SPI and other 2 on SPI1.  Can update to make more
compatible with systems that don't have SPI1.

Some of the modes:
a) Use the currently released api to read each of these
b) Use new one that starts up range on all 4 devices and waits for them
to complete by doing query to see if complete.
c) Like b) but instead checks GPIO pin for completion.
d) Continuous updates - allows you to type in cycle time.

Again there is lots more testing and extensions that can and should be
done.

Note: I ran clan-format -i on the three files.