examples  Multi_extended work without Wire1 plus low memory
Not all platforms have Wire1... So detect it first and if platform does not have it, use Wire instead.

Also added #defines at top to allow easier to change

Reduce the number of objects on AVR based boards.

And changed several of the messages to use program memory instead of
real memory